requirements.txt
opencv-python
tensorflow
keras
numpy
Pillow
ttkthemes
